Copy edit
In the Hardware list, it specifies bell wire, but this isn't mentioned again, and this worksheet suggests using 2 core speaker cable. I'm aware that bell wire is cheaper, but speaker cable is possibly more durable. Can we clarify which we want the user to buy, so either this worksheet or the hardware page can be changed accordingly?